server: add tls module
